application/requirements.txt
confluent_kafka==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.25.1
validators==0.28.3
